+++ README.md
@@ -1,20 +1,33 @@
-# Cheater Dice Protector
+# Dice RNG Protector
 ## [Foundry VTT](https://foundryvtt.com) Module
 ---
-It's easy to cheat your roll in Foundry with a little bit of Javascript knowledge. This module block the most commun way to alterate the dice result.
+It's easy to cheat your roll in Foundry with a little bit of JavaScript knowledge. This module block the most common way to change the dice result.
 
 ## Changelog
 
+### 0.1.4
+Migration to v10
+
+Additions:
+* Spanish translation.
+* GitHub workflow for releases.
+
+Changes:
+* Tweaked English and French translations.
+* Changed RNG function for more "random" results.
+* Changed folder structure to match the v10 template
+* Many other minor changes.
+
 ### 0.1.3
-* Add compatibility for version 9
-* Add Fudge feature
+* Add compatibility for version 9.
+* Add Fudge feature.
 
 ### 0.1.2
-* Convert french file ANSI to UTF8
+* Convert french file ANSI to UTF8.
 
 ### 0.1.1
+* Add setting to activate protection only for players.
 * Add setting to choose the number generator you want.
-* Add setting to activate protection only for players
 
 ### 0.1.0
-* Initial release, freeze the number genrator
+* Initial release, freeze the number generator.
