--- lang/en.json
@@ -1,23 +1,0 @@
-{
-  "dcp": {
-    "Normal": "Normal",
-    "Minimal": "Min",
-    "Maximal": "Max",
-    "Low": "Low",
-    "High": "High",
-    "settings": {
-      "usecrypto": {
-        "name": "Use Web Crypto API",
-        "hint": "Use the modern cryptography module available in browsers instead of Mersenne Twister (1997) for dice results."
-      },
-      "onlyforplayer": {
-        "name": "Only for player",
-        "hint": "Protect the dice random generator only for players"
-      },
-      "enablefudgedice": {
-        "name": "Enable Fudge",
-        "hint": "DM can modify his 'luck' (refresh require)"
-      }
-    }
-  }
-}
