--- lang/fr.json
@@ -1,23 +1,0 @@
-{
-  "dcp": {
-    "Normal": "Normal",
-    "Minimal": "Min",
-    "Maximal": "Max",
-    "Low": "Bas",
-    "High": "Haut",
-    "settings": {
-      "usecrypto": {
-        "name": "Utiliser Web Crypto API",
-        "hint": "Utiliser le module de cryptographie du navigateur internet au lieu de Mersenne Twister (1997) pour les jets de dés"
-      },
-      "onlyforplayer": {
-        "name": "Seulement pour les joueurs",
-        "hint": " Protège le générateur de nombre aléatoire pour les joueurs seulement"
-      },
-      "enablefudgedice": {
-        "name": "Modifier les dées du MJ",
-        "hint": "Le MJ peut modifier sa 'chance' (rédémarrage requis)"
-      }
-    }
-  }
-}
